add SNP allele strand flippin to AncestryDNA filetype
AncestryDNA data download exports all results on FWD strand, not the standard strand  (which may be FWD or REV depending on the SNP).  23andMe uses standard strand, as does snpedia.  So from now on all AncestryDNA uploads will be converted to standard strand convention on upload, and thereafter we will not have to worry about orientation since it is consistent for all uses.